Improve database performance
This improves database performance by adding generated-columns to the "player_chunk" table so that "getChunkTimestamps()" does not need to generate a disposable ghost table with these columns on-demand. This halves the time it takes to execute the query.

This is done through a reversible migration system, so existing databases (like that example 4.6GB database) can be gracefully updated.